requirements.txt
accelerate==0.26.1
antlr4-python3-runtime==4.9.3
certifi==2023.11.17
charset-normalizer==3.3.2
diffusers==0.24.0
filelock @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_d3quwmvouf/croot/filelock_1700591194006/work
fire==0.5.0
fsspec==2023.12.2
ftfy==6.1.3
gmpy2 @ file:///Users/cbousseau/work/recipes/ci_py311/gmpy2_1677937751357/work
huggingface-hub==0.20.2
idna==3.6
importlib-metadata==7.0.1
Jinja2 @ file:///Users/cbousseau/work/recipes/ci_py311/jinja2_1677916113134/work
MarkupSafe @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_a84ni4pci8/croot/markupsafe_1704206002077/work
mpmath @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_17iu6a8a3m/croot/mpmath_1690848269369/work
networkx @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b9af3smw_7/croot/networkx_1690562010704/work
numpy==1.26.3
nvidia-pyindex==1.0.9
omegaconf==2.3.0
onnx==1.15.0
onnx-graphsurgeon==0.3.27
packaging==23.2
pillow==10.2.0
polygraphy==0.49.0
protobuf==4.25.2
psutil==5.9.7
python-dotenv==1.0.0
PyYAML @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_a8_sdgulmz/croot/pyyaml_1698096054705/work
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scipy==1.11.4
six==1.16.0
streamdiffusion==0.1.1
sympy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_7cbpg8656h/croot/sympy_1701397648473/work
termcolor==2.4.0
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
typing_extensions @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_99x3ghmw54/croot/typing_extensions_1705005641392/work
urllib3==2.1.0
wcwidth==0.2.13
zipp==3.17.0
